Remove ownerState
